+++ cli/js/net.ts
@@ -1,6 +1,6 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 import { EOF, Reader, Writer, Closer } from "./io.ts";
-import { read, write } from "./files.ts";
+import { read, write } from "./ops/io.ts";
 import { close } from "./ops/resources.ts";
 import { sendSync, sendAsync } from "./ops/dispatch_json.ts";
